\name{bindata}
\alias{bindata}
\docType{data}
\title{ Simulated binary data }
\description{
Simulated data consisting of 24 binary variables and a binary outcome \code{Y} with 100 observations. Names of variables associated with the outcome start with \code{true}, names of other variables start with \code{zero}. 
}

\usage{data(bindata)}
%\format{}
%\references{}
\seealso{\code{\link{gGlobalAncova}}}
\examples{
data(bindata)
#str(bindata)
}
\keyword{datasets}
