\name{group.info}
\alias{group.info}
\docType{data}
\title{Prognostic stages of patients  }
\description{
 Group index for the p53 example:
 Data on the prognostic stage of 36 patients with colorectal cancer
 \code{0} - UICC II
 \code{1} - UICC III
 }
\usage{data(group.info)}
\format{
  The format is:
 num [1:36] 0 1 1 0 0 1 0 1 0 0 ...
 }
\examples{
data(group.info)
#table(group.info)
}
\keyword{datasets}

